requirements.txt
asgiref==3.4.1
Django==3.2.7
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
flake8==3.9.2
mccabe==0.6.1
mysqlclient==2.0.3
phonenumbers==8.12.31
Pillow==8.3.2
pycodestyle==2.7.0
pyflakes==2.3.1
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.2
